requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.0
tzdata==2021.5
whitenoise==5.3.0
